# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Survey API",
        "description": "REST API сервиса опросов и голосований",
        "contact": {
            "name": "API Support",
            "email": "support@survey.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/surveys/{survey}/questions": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Создать вопрос в опросе (только автор, черновик)",
                "operationId": "9513064f0c0ae507465b7400c9ca31e7",
                "parameters": [
                    {
                        "name": "survey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text",
                                    "type",
                                    "order"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Какой ваш любимый язык программирования?"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "single",
                                            "multiple",
                                            "text"
                                        ],
                                        "example": "single"
                                    },
                                    "order": {
                                        "description": "Порядок вопроса",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Вопрос создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Это не ваш опрос"
                    },
                    "422": {
                        "description": "Нельзя добавлять вопросы к опубликованному опросу"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/surveys/{survey}/questions/{question}": {
            "put": {
                "tags": [
                    "Questions"
                ],
                "summary": "Обновить вопрос (только автор, черновик)",
                "operationId": "16b4db88f714988136bf1c3ed6f05f1e",
                "parameters": [
                    {
                        "name": "survey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "question",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "Обновленный текст вопроса"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "single",
                                            "multiple",
                                            "text"
                                        ]
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Вопрос обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Question"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Это не ваш опрос"
                    },
                    "404": {
                        "description": "Вопрос не принадлежит опросу"
                    },
                    "422": {
                        "description": "Нельзя редактировать вопросы опубликованного опроса"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Удалить вопрос (только автор, черновик)",
                "operationId": "069ffbdac506f923338ac7d6828f1b95",
                "parameters": [
                    {
                        "name": "survey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "question",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Вопрос удален",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Это не ваш опрос"
                    },
                    "422": {
                        "description": "Нельзя удалять вопросы опубликованного опроса"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/surveys/{survey}/responses": {
            "post": {
                "tags": [
                    "Responses"
                ],
                "summary": "Пройти опрос — отправить ответы",
                "operationId": "eacfeae95bef48871d25c55a335299b4",
                "parameters": [
                    {
                        "name": "survey",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "option_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    },
                                                    "example": [
                                                        1
                                                    ]
                                                },
                                                "text_value": {
                                                    "type": "string",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ответы приняты"
                    },
                    "422": {
                        "description": "Ошибка валидации или повтор"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Question": {
                "title": "Вопрос",
                "description": "Вопрос в опросе",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "survey_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "Какой ваш любимый язык?"
                    },
                    "type": {
                        "description": "Тип вопроса: single (одиночный выбор), multiple (множественный выбор), text (текстовый)",
                        "type": "string",
                        "enum": [
                            "single",
                            "multiple",
                            "text"
                        ],
                        "example": "single"
                    },
                    "order": {
                        "description": "Порядок вопроса",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "options": {
                        "description": "Варианты ответов (только для single/multiple)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        }
                    }
                },
                "type": "object"
            },
            "Option": {
                "title": "Вариант ответа",
                "description": "Вариант ответа на вопрос",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "question_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "PHP"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT токен в формате: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}